requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
streamlit
requests
apscheduler
vonage
psycopg2
python-dotenv